Segregate promise code. Author race
